Refactor: Moved our python-embedder into a more correct package structure
