+++ python-embedder/app/routes/__init__.py
@@ -1,0 +1,2 @@
+from .embedding import bp as embedding_bp  # noqa: F401
+from .eval import bp as evaluation_bp  # noqa: F401
